+++ libraries/fdr/include/mcrl2/fdr/detail/utility.h
@@ -45,7 +45,7 @@ bool check_term_argument(Term t, CheckFunction f)
 // 2) if the list has the proper minimum size
 // 3) if all elements of the list satisfy the predicate f
 template <typename Term, typename CheckFunction>
-bool check_list_argument(Term t, CheckFunction f, unsigned int minimum_size)
+bool check_list_argument(Term t, CheckFunction f, size_t minimum_size)
 {
   atermpp::aterm term(atermpp::aterm_traits<Term>::term(t));
   if (t.type() != AT_LIST)
